+++ NEWS.md
@@ -1,3 +1,7 @@
-# dilp (development version)
-
+# dilp (development version 0.0.0.9001)
 * Initial CRAN submission.
+
+# dilp (development version 0.0.0.9000)
+
+* Functions implemented for digital leaf physiognomy (dilp())
+* Functions implemented for leaf mass per area reconstruction (lma())